higan: synchro scheduler
Adding an alternative scheduler which reduces the need for
co_switch to two per frame. This required for Thread to define
an abstract main method (which nearly all inheriting classes already had),
and to make alternative implementations of main where Thread::synchronize
is always at the tail-end of an execution branch. This should
normally allow for as much cross-compatibility between the standard
scheduler and the alternative one.